refactor(core): 🔨 split long SearchEvents methods
